- mdmkolbe fixed some minor warnings.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@25109 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ apps/tools/genmeshify/converter.cpp
@@ -704,7 +704,7 @@ namespace genmeshify
       if (!WritePolyMesh (key, polymeshNode)) return false;
 
       static const char* const meaningKeys[] = {"colldet", "shadows", "viscull" };
-      for (int i = 0; i < sizeof (meaningKeys)/sizeof (meaningKeys[0]); i++)
+      for (size_t i = 0; i < sizeof (meaningKeys)/sizeof (meaningKeys[0]); i++)
       {
         if (meaning.Check (1 << i))
         {
